requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
apipkg==1.5
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
execnet==1.7.1
glob2==0.7
idna==2.9
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
parse==1.15.0
parse-type==0.5.2
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-bdd==3.3.0
pytest-forked==1.1.3
pytest-xdist==1.32.0
requests==2.23.0
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
